Refactor out run_eval out of generate.py
